% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_new.R
\name{DataPreprocess}
\alias{DataPreprocess}
\title{Preprocess the data}
\usage{
DataPreprocess(
  sce,
  n.PCs = 50,
  norm.type = c("logOnly", "logNorm"),
  select.hvg = FALSE,
  n.HVGs = 2000,
  assay.type = "logcounts",
  BSPARAM = ExactParam()
)
}
\arguments{
\item{sce}{SingleCellExperiment to preprocess}

\item{n.PCs}{Number of principal components to compute. Default is 50.}

\item{norm.type}{Type of normalization. logOnly: only take logarithm.
logNorm: compute log-transformed normalized expression values from a count matrix.}

\item{select.hvg}{Select highly variable genes to run PCA upon. Default is FALSE.}

\item{n.HVGs}{Number of highly variable genes. Required if select.hvg is TRUE. Default is 2000.}

\item{assay.type}{Name of assay in \code{sce} containing normalized counts.
Leave as "logcounts" unless you explicitly pre-computed a different
normalization and added it to \code{sce} under another assay.}

\item{BSPARAM}{A \linkS4class{BiocSingularParam} object specifying which
algorithm should be used to perform the PCA. By default, an exact PCA is
performed, as current spatial datasets are generally small (<10,000 spots).
To perform a faster approximate PCA, please specify
\code{FastAutoParam()} and set a random seed to ensure
reproducibility.}
}
\value{
SingleCellExperiment with PCA
}
\description{
This code is based on the R function "spatialPreprocess" in BayesSpace
}
