% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_new.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Data: example_data}
\usage{
data(example_data)
}
\description{
example_data contains coord and gene_data_pc
coord: coordinates of the cells
gene_data_pc: gene expression data after PCA
}
