% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_new.R
\name{FindFiniteSet_c_R}
\alias{FindFiniteSet_c_R}
\title{Determine the finite set {k: pi_k >= u_i} for each spot i}
\usage{
FindFiniteSet_c_R(u, K_max, dpAlpha, a_eta, b_eta, dpXi, eta_k)
}
\arguments{
\item{u}{Augmented variable for spot i.}

\item{K_max}{Maximum in the cell type indicators C.}

\item{dpAlpha}{Concentration parameter in Dirichlet process.}

\item{a_eta}{Mean of the normal prior for \eqn{\eta}.}

\item{b_eta}{Standard deviation of the normal prior for \eqn{\eta}.}

\item{dpXi}{Beta random variables in the stick-breaking process.}

\item{eta_k}{Mean value matrix for gene expression data. Rows: cell type clustering number. Columns: genes.}
}
\value{
FindFiniteSet_c_R returns an R list including the following information.
\item{finiteSet}{vector, the finite set this function aims to find.}
\item{dpXi}{vector, beta random variables in the stick-breaking process.}
\item{eta_k}{matrix, updated mean value matrix (maybe add some samples drawn from prior).}
\item{K_max}{integer, updated value of K_max (after possibly adding some samples).}
}
\description{
Determine the finite set {k: pi_k >= u_i} for each spot i
}
